+++ include/Tact/Spatializer.hpp
@@ -8,7 +8,7 @@ namespace tact
 class SYNTACTS_API Spatializer {
 public:
     struct Point {
-        float x,y;
+        double x,y;
     };
     Spatializer(Session* session = nullptr);
     ~Spatializer();
@@ -16,15 +16,15 @@ public:
     void bind(Session* session);
     void unbind();
 
-    void setPosition(int channel, float x, float y = 0);
+    void setPosition(int channel, double x, double y = 0);
     void setPosition(int channel, Point p);
     Point getPosition(int channel) const;
 
-    void setTarget(float x, float y = 0);
+    void setTarget(double x, double y = 0);
     void setTarget(Point p);
     Point getTarget() const;
-    void setRadius(float r);
-    float getRadius() const;
+    void setRadius(double r);
+    double getRadius() const;
     void setRollOff(Curve rollOff);
     Curve getRollOff() const;
 
@@ -36,19 +36,18 @@ public:
 
     void play(Signal signal);   
     void stop();
-    void setVolume(float volume);
-    void setPitch(float pitch);
+    void setVolume(double volume);
+    void setPitch(double pitch);
 
     void save(const std::string& name);
     void load(const std::string& name);
 private:
     void update();
 private:
-
     Session* m_session;
     Point m_target;
-    float m_radius;
-    float m_volume;
+    double m_radius;
+    double m_volume;
     Curve m_rollOff;
     std::map<int,Point> m_positions;
 };
