+++ src/Tact/Curve.cpp
@@ -6,48 +6,48 @@ namespace tact
 
 Curve::Curve() : Curve(Curves::Linear()) {}
 
-float Curve::operator()(float t) const
+double Curve::operator()(double t) const
 {
     return m_ptr->operator()(t);
 }
 
-float Curve::operator()(float a, float b, float t) const
+double Curve::operator()(double a, double b, double t) const
 {
     return lerp(a, b, m_ptr->operator()(t));
 }
 
 namespace Curves
 {
-float Instant::operator()(float t) const
+double Instant::operator()(double t) const
 {
     return 1;
 }
 
-float Delayed::operator()(float t) const
+double Delayed::operator()(double t) const
 {
     if (t < 1)
         return 0;
     return 1;
 }
 
-float Linear::operator()(float t) const
+double Linear::operator()(double t) const
 {
     return t;
 }
 
-float Smoothstep::operator()(float t) const
+double Smoothstep::operator()(double t) const
 {
     t = t * t * (3.0f - 2.0f * t);
     return t;
 }
 
-float Smootherstep::operator()(float t) const
+double Smootherstep::operator()(double t) const
 {
     t = t * t * t * (t * (t * 6.0f - 15.0f) + 10.0f);
     return t;
 }
 
-float Smootheststep::operator()(float t) const
+double Smootheststep::operator()(double t) const
 {
     t = t * t * t * t * (t * (t * (t * -20.0f + 70.0f) - 84.0f) + 35.0f);
     return t;
@@ -60,19 +60,19 @@ float Smootheststep::operator()(float t) const
 namespace Quadratic
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = t * t;
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t = t * (2.0f - t);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -90,20 +90,20 @@ float InOut::operator()(float t) const
 namespace Cubic
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = t * t * t;
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t -= 1.0f;
     t = (t * t * t + 1.0f);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -121,20 +121,20 @@ float InOut::operator()(float t) const
 namespace Quartic
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = t * t * t * t;
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t -= 1.0f;
     t = -(t * t * t * t - 1.0f);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -152,20 +152,20 @@ float InOut::operator()(float t) const
 namespace Quintic
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = t * t * t * t * t;
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t -= 1.0f;
     t = (t * t * t * t * t + 1.0f);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -187,19 +187,19 @@ float InOut::operator()(float t) const
 namespace Sinusoidal
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = 1.0f - std::cos(t * HALF_PI);
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t = std::sin(t * HALF_PI);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t = -0.5f * (std::cos(PI * t) - 1.0f);
     return t;
@@ -210,19 +210,19 @@ float InOut::operator()(float t) const
 namespace Exponential
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = std::pow(2.0f, 10.0f * (t - 1.0f));
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t = (-std::pow(2.0f, -10.0f * t) + 1.0f);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -240,20 +240,20 @@ float InOut::operator()(float t) const
 namespace Circular
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = (1.0f - std::sqrt(1.0f - t * t));
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t -= 1.0f;
     t = std::sqrt(1.0f - t * t);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -275,7 +275,7 @@ float InOut::operator()(float t) const
 namespace Elastic
 {
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     if (t == 0 || t == 1)
         return t;
@@ -284,14 +284,14 @@ float In::operator()(float t) const
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     if (t == 0 || t == 1)
         return t;
     return (std::pow(2.0f, -10.0f * t) * std::sin((t - 0.1f) * (2.0f * PI) * 2.5f) + 1.0f);
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -312,23 +312,23 @@ float InOut::operator()(float t) const
 namespace Back
 {
 
-constexpr float s = 1.70158f;
-constexpr float s2 = 2.5949095f;
+constexpr double s = 1.70158f;
+constexpr double s2 = 2.5949095f;
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     t = (t * t * ((s + 1.0f) * t - s));
     return t;
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     t -= 1.0f;
     t = (t * t * ((s + 1.0f) * t + s) + 1.0f);
     return t;
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     t *= 2.0f;
     if (t < 1.0f)
@@ -350,7 +350,7 @@ namespace Bounce
 
 namespace detail
 {
-float Out(float t)
+double Out(double t)
 {
     if (t < (1.0f / 2.75f))
     {
@@ -373,12 +373,12 @@ float Out(float t)
     }
 }
 
-float In(float t)
+double In(double t)
 {
     return 1.0f - Out(1.0f - t);
 }
 
-float InOut(float t)
+double InOut(double t)
 {
     if (t < 0.5f)
         return In(t * 2.0f) * 0.5f;
@@ -386,17 +386,17 @@ float InOut(float t)
 }
 } // namespace detail
 
-float In::operator()(float t) const
+double In::operator()(double t) const
 {
     return detail::In(t);
 }
 
-float Out::operator()(float t) const
+double Out::operator()(double t) const
 {
     return detail::Out(t);
 }
 
-float InOut::operator()(float t) const
+double InOut::operator()(double t) const
 {
     return detail::InOut(t);
 }
